pyproject.toml
[tool.poetry]
name = "little-learners-tech-backend"
version = "0.1.0"
description = "Django backend for Little Learners Tech platform"
authors = ["Little Learners Tech <contact@littlelearners.tech>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.0"
djangorestframework = "^3.14.0"
firebase-admin = "^6.2.0"
google-cloud-secret-manager = "^2.16.4"
python-decouple = "^3.8"
django-cors-headers = "^4.3.0"
psycopg2-binary = "^2.9.0"
django-extensions = "^3.2.0"
python-dotenv = "^1.0.0"
dj-database-url = "^2.1.0"
pillow = "^11.3.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
stripe = "^12.5.1"
coreapi = "^2.3.3"
slack-sdk = "^3.36.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest-django = "^4.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
